ARUHA-49 Event Type registration
Creates a Kafka Topic together with Event Type persistence, since there
is a 1:1 relation between then.

If the topic creation fails, rollback Event Type persistence.

This implementation assumes that the Aplication is not going to
terminate inexpectedly.

But if it does sudenly terminates it might let the system in an
incosistent state, where we have a database entry for the Event Type but
no Kafta Topic.

It would require a different approach in order to achieve guaranteed consistency.