aruha-104 event type in memory cache
Caching event types in memory could significantly improve the
performance of event publishing, since we need to load event types every
time to check for the schema.

We decided to use an eventual consistent cache synchronization, which
means that it might take some time to propagate changes to all nakadi
instances. This solution was chosen over fully synchronized one due to
CAP (performance) and it fits well the most usual use cases.