aruha-104 remove test that test too few
In order for this test to really test something, it would be required to
first erase the znode, then instantiate a cache and observe that it
created all the required nodes. But removing this znode has side effects
in the user journey test since it disables the app cache.

Once we had separated integration tests from AT we could restore this
test.