aruha-108 remove unused prototyp validation code
The EventValidation class was supposed to work a repository of event
type validators but this code is not used in the project, since we have
a guava chace in place.

Guava chache provides more guarantees for generating event type
validators one and one time only. The current code, even if thread safe,
could lead to multiple event types been created because it's not
synchronized.