Based on commitid 65bb653d3 and must be used for live deployment of
eventstore if it is necessary. In addition to that I've added shell
script which starts up gunicorn. Amount of gunicorn workers can be
specified as an environment variable `GUNICORN_WORKERS`. If this value
is not specified, it will calculate it from the amount of cpu cores with
using following formula: 2 * nproc + 1