aruha-112 use ack all property for durable messages
The default kafka producer acknowledgement property is "1". As by the
documentation:

"acks=1 This will mean the leader will write the record to its local log
but will respond without awaiting full acknowledgement from all
followers. In this case should the leader fail immediately after
acknowledging the record but before the followers have replicated it
then the record will be lost."

E.g. we would reply to the user that a message was successfully
published but eventually it would fail to replicate and that means it
would never be available for consumption. So the message would be lost.