aruha-112 repository method for batch delivery
This commit introduces a new method for sync delivery of event batches.

Kafka producer tries to batch by default, so the only change from the
single event delivery to the batch delivery is that in the batch we
"queue" many events and then flush the all batches in order to wait for
futures to resolve, this way we can properly update the status of each
message delivery.

In order to represent delivery status we introduced some new domain
classes.