aruha-112 remove in memory implementations
The functionality provided by repositories is growing in size and
complexity therefore increasing the cost of maintainance for in memory
implementations.

Besides that, the in memory implementation no longer simulates properly
the actual kafka and db implementations, which could lead to false
positives while implementing tests.

We should favor mocks in detriment of in memory, even though it can be
 expensive to maintain as well, at least we have more control and can
 test every possible scenario, while in memory usually only allows for
 happy path tests.