aruha-112 create bulk submission procedure
This service class is responsible for submitting a batch of events to
the underlying message service (kafka).

The choice for returning a result object instead of throwing exceptions
is because we needed to return the status for each item instead of a
global status only.

We could have added a field in the exception to hold such list of status
but I judged that a result object would better express the semantics of
a call to this method and further help users to write client code that
handles every scenario: SUBMITED, ABORTED, FAILED.

This class also is a domain level class and contain no information about
the HTTP layer, which is one of the design goals.