aruha-112 change single submit to batch submit
This commit introduces a backward incompatible change to Nakadi
API. Every event submission is now a bulk submission.

So if users need to submit single events, they should be wrapped in an
array with a single element.

Since we moved the publishing logic to a service class, we made the
controller implementation simpler by just calling this procedure. The
tests are also simpler since the controller now has fewer
responsabilites.