aruha-131 validate partition strategy on update
The partition strategy was not being validated on the update. So would
be possible to do stuff with PUT that is not possible on the POST. For
example, defining hash partition strategy without a
patition_key_fields. But this is not possible on POST.

This could cause 500 when publishing events to an event type with
inconsistent partition strategy data.