aruha-68 enrichment
Applies enrichment to events. It mutates the original event object in
order not to create to many objects and keep things as fast as possible.

We only have one enrichment strategy at the momemnt, but it's done in a
way that would be possible to add new strategies or even change the
registry implementation to a dynamic one.