aruha-151 fix definitions on data change events
Since we wrapping the original json schema inside a wrapper schema,
references to definitions in the root of the provided JSON were broken.

This commit fixes this issue by moving definitions to the root of the
wrapper json.

If there are other definitions at other levels of the provided json
schema, that would stil be an unsolved issue.