ARUHA-240: Update Event Type name description.
This removes the policy recommendations and requirements for
event type owner naming, and replaces it with a note saying
the name is regex constrained and Nakadi users may decide to
apply a naming policy. The API itself doesn't support encoding
ownership in names, and whether to do so is a local decision for
people running and using Nakadi.