aruha-246 fix data change event partition
When defining `partition_key_fields` users expect that the lookup path
to be relative to the originally defined schema, but since Nakadi
manipulates the schema for DataChangeEvent's, we need to add this
special case in order to scope the path to the nested "data" field.

This ways users no longer need to prepend "data." to
partition_key_fields.