aruha-190 posting cursors as an extensible object
We are switching the post endpoint to an object instead of an array in
order to make it easear to extend in the future without introducing
backward incompatibilities.

Also, moving nakadi stream from the Header to a property in this
object. We are going to favor this URI + entity approach instead of
relying on headers a lot, whenever possible. That is usually more
familiar to users.