aruha-190 remove commit_timeout and add StreamInfo
Allowing for clients to control commit_timeout is a dangerous thing,
since it reserves resources in the server. We are removing this option
in order to make operations easier and provide more control to Nakadi to
close connections when it judges necessary.

On the other hand, we are introducing a new object to the
EventStreamBatch. We believe it will help debug issues and make the
learning curve to implement clients smoother.