aruha-338 rename window_size and change default value
Renaming window_size to max_uncommitted_size makes the API easier to
understand since it doesn't introduce any new terminology (window
concept is not present anywhere in the api so far).

We are also reducing the default amount to 10 since there are some
potentialy large messages and a default of 100 would mean to stream
several megabytes at once.