aruha-338 remove commit_timeout
Allowing clients to control the commit_timeout implied in potentially
illimited resource allocation in the server, leading to security and
operational problem.

We removed this parameter from the API but keep it as an application
configuration.

The default value was also changed to 60 seconds for lauching locally
but as 5 seconds for testing purposes.