aruha-190 remove id from cursors
The original idea for including an `id` property to cursors were to make
then identifiable, so to easy the task of
debugging (https://github.com/zalando/nakadi/pull/353#discussion_r76253962).

Although, up to now, no state is stored when delivering cursors to
clients. E.g. Nakadi doesn't keep track of what cursors were delivered.

If we are going to identify cursors with a uuid, in order to recognize
such ID later, when a consumer commits it, it would be required to hold
such information in the server.

Since making the server statefull is not an option, we are removing this
field from the API definition. We believe that improving tracebility and
making debugging easy can be achieved using other means, as for example,
the newly added field cursor_token.