aruha-190 define offsets format
Up to now nakadi Cursors were advertised to be opaque entities and
clients should not make assumptions about it's values.

But we've been observing clients making assumptions and coding based on
them. This could lead to a major confusion in the future if we no longer
hold those assumptions true.

In order to mitigate this risk, we are committing to a well defined
structure beforehand. It's inspired in the postgres WAL format
https://www.postgresql.org/docs/9.2/static/wal-internals.html and we aim
to address two issues with this format:

1. Make Nakadi operations easier (this feature would have been
tremendously useful for cluster migration or implementing partition
rebalacing)
2. Provide a well defined set of guarantees for consumers to rely one.