aruha-338 add metadata debugging info to first and last batch
In order to provide more information for users debug their
implementation, we are writting the reason for closing a connection in
the last batch.

We could introduce more information to this stream metadata in the
future, so that clients would be notified of important events such as
rebalances or even help then distinguish between empty logs or blocking
commits.

To make sure that clients won't be suprised by such field at the very
last batch, we are also sending a welcome message in the fist batch, so
that if someone is implementing in a fragile way, their code will break
early so that can fix.