aruha-338 avoid code reentrance
Previous to this commit, there could be a problem when shutting down
gracefully due to the client closing the connection. It would enter a
loop that would not allow the server to immediately close the connection
from its side, as well.

Decoupling from shutdown gracefully and sending messages solved the
problem. We are only sending the metadata when nakadi thinks the
connection is still open.