aruha-384 reduce the maximum number of database connections
We had a fuckup due to the number of database connections increasing the
limit. But we only use the database to load event types for caching.

For this reason, we are going to reduce the amount of connections, so to
be able to increase the number of Nakadi instances without using all
database connections.