aruha-116 clarify on forbidden attributes
Following the design of open for extension, closed for changes, only
"not" and "patternProperties" are not allowed since they don't fit under
this concept. Negation requires the definition of extensibility in the
negation context which is counter intuitive. Pattern properties is hard
to define since it would close an object for further extension
inadvertently.