aruha-400 prepare for uknown attributes
In order to introduce some new fields to event types (compatibility mode
and schema versioning), I plan to first run a migration that would do
so.

But once we run this database migration that adds such fields to event
types, it would break the running version, since it would introduce new
fields that are not know to the current stack.

In order to avoid this incompatibilities, I will first release this
feature firts, which ignores uknown attributes instead of breaking.