aruha-400 refactor schema constraint check
This refactoring uses a json schema definition of what a Nakadi json
schema is allowed to use. It follows a declarative approach and relies
on well documented and tested everity library to do the job correctly.

One trade off is the error messages might not be as clear as they were
previously. But I considered them to be understandable and correct.