aruha-578 Add changelog
In order to better communicate changes we are going to maintain a change
log file following a well defined standard.

Since this standard required semver adoption, we are using this
opportunity to bump version from 0.8.0 to 1.0.0, which better reflects
the maturity of the project's API definition.

Read more about the standard at http://keepachangelog.com