aruha-584 consumer quota
In order to use Kafka quota on a fine grained level, we are going to
identify consumers using a combination of application name and event
type name. This will guarantee that an application can stream up to the
default quota from a single partition, since Kafka quotas are
implemented per client-id per broker.