ARUHA-641 Write batch events directly to outputstream.
If SEND_BATCH_VIA_OUTPUT_STREAM is enabled, send the batch and its events
directly to the consumer's OutputStream as bytes instead of building up an
intermediate batch String via createStreamEvent. The standard parts of the
batch are allocated statically as bytes instead of per call.

This is done to reduce the number of allocations generated for a batch and is
based on some performance tuning investigations.