Write batch subscription events directly to the output stream.
If SEND_SUBSCRIPTION_BATCH_VIA_OUTPUT_STREAM is enabled, send the batch and
its events directly to the SubscriptionOutput's underlying OutputStream as
bytes instead of building up an intermediate batch String via serializeBatch.
The standard parts of the batch are allocated statically as bytes instead of
per call. This is done to reduce the number of allocations generated for a
batch and is based on some performance tuning investigations.

Each call to SubscriptionOutput.streamData calls flush on its underlying
OutputStream. This might be a problem and if it is, SubscriptionOutput will
probably have to allow StreamState to access to its OutputStream.