aruha-541 consumer lag monitoring endpoint
Previous to Timelines, cursor's offsets used to be sequential Long
numbers as Strings, for example "00000123". Because of that, several
consumers used to convert such strings back into numbers, our example
above would be converted to 123, and later use them as numbers for
several purposes.

One of the things consumers used to do was to query for `/partitions`
endpoint to get the newest available offset and then calculate the
distance between the newest and the latest consumed offset. This is how
many consumers from low level API would detect lag in consumer.

After introducing timelines, it will be no longer possible to easily
convert an offset to a number and perform such operations on the client
side. For that reason, we are introducing an extension to the
`/partitions` endpoint that will allow consumers to get the number of
unconsumed events without the need to perform any conversions or
calculations on their side.