aruha-691 Cursor navigation
Previous to Timelines it was possible to take one offset, for example
"000000000000000009", transform it into a number and later add or
subtract a given number from it, in order to craft another valid offset
that could be used for random access.

After timelines, cursors are no longer simple sequences and can no
longer be easily or safely manipulated by the client for random access.

For this reason we are providing this extension to the API, as an
alternative for randomly accessing the stream.