ARUHA-692 Cursor comparison
Previous to Timelines it used to be possible to calculate how many
events here is between two offsets. For example, between offsets
"000000000000000001" and "000000000000000005" there are obviously 4
events.

This kind of comparison is used extensively by consumers in order to
verify that their implementation is correctly processing the same number
of events that Nakadi sent to them.

This endpoint provides an alternative way of calculating this distance
after Timelines was introduced.