aruha-541 change X-nakadi-cursors from header to query parameter
The usage of header has been discouraged by the Api Guild for a number
of reasons https://github.com/zalando/restful-api-guidelines/issues/207

As an alternative, we are going to encode this parameter in the URL,
since we are not worried about reaching the mithical maximum size
limit of 2 KB chars.

This commit also includes further clarifications on this parameter
requirements and behaviour.