aruha-531 cursor operator refactoring
Delegate offset conversion from string to number to the specific topic
repository. It means that it's repository's responsability to do such
conversions, which makes the service more generic. I.e. it should
automatically work with any type of repository that implements the 3 new
added methods.

It already prepared for the addition of Kinesis, for example.