aruha-531 clean up parameter validation
Previous validation wasn't working at all. When using @Valid List as a
paremeter, it doesn't validate the items in the list since List is not a
JavaBean.

The solution I chose was to wrap the list in a bean in order to validate
it.

Now it works. This change should be applied to the other places that
expect arrays as parameters, if any.