aruha-677 force all event types to use proper json-schema
This seems to be a very simple change but it's a very dramatic one.

By default, json-schema is an extensible specification. It means that
user can be very inventive in the attributes they want to define, even
though they are not in the specification. For example, some users like
to provide "example" attribute with some example value. This is not in
the specification, but validator usually just don't care about it, since
it has no meaning.

The problem is that sometimes, readers of these schemas are not sure if
such fields have any meaning or not. In order to avoid this
misunderstandings, we are going to enforce all schemas to contain only
attributes that are in the v4 draft of json-schema specification.

Anything else will result in failure.