Document timeline creation process
The process of creating a Timeline involves coordination of all Nakadi
instances using many different concurrent programming techniques:
barriers, locks and async notifications.

It's a non trivial process and for that reason we wrote how it works in
English. We believe that it might help new developers to understand how
it works and also help with knowledge sharing.

It focuses mostly on the structures used and not the actual code, which
developer should be able to read without any problem.