aruha-778 Set login, connection, and socket timeouts for postgres
Set timeouts for postgres, to avoid connections hanging forever if
Postgres is restarted. All three timeouts are set to 2 seconds. Here is
the relevant section from the postgres jdbc documentation:

loginTimeout = int
Specify how long to wait for establishment of a database connection.
The timeout is specified in seconds.

connectTimeout = int
The timeout value used for socket connect operations. If connecting to
the server takes longer than this value, the connection is broken.
The timeout is specified in seconds and a value of zero means that it
is disabled.

socketTimeout = int
The timeout value used for socket read operations. If reading from the
server takes longer than this value, the connection is closed. This can
be used as both a brute force global query timeout and a method of
detecting network problems. The timeout is specified in seconds and a
value of zero means that it is disabled.