aruha-840 optimize subscription streaming
This PR is a "rebase" of #590.

Since a lot has changed in the code, it's not a true rebase, but an
adaptation. The new streaming code no longer uses a `List<String>`
events but instead a `List<ConsumedEvent>`.

In order to avoid transforming the later into the former, and allocating
a new list on every call to write to output stream, I created a new
method writeSubscriptionBatch that accepts lists of `ConsumedEvents`.

This code is not the most straighforward, but the intention is to avoid
allocating extra lists and objects. If you have a better idea on how to
implement it, please, share it.