Do not override flow_id when present
Nakadi used to expect the flow_id to be taken from a header but not even
a single user was happy that their flow_id has been
overriten. E.g. users do not expect their event to be changed if they
are sending something already.

It's also a general guideline from Nakadi, not to "enrich" events,
always favoring user defined data. We do not change users data and this
bug broke this commitment.