aruha-929 unify resource representing an event type
There were two different resources for representing an event type. I
chose the one that had fewer parameters in the constructor and offered
the same functionality. It also makes sure that `getType` alwyas results
in the same string.

But I used the name `EventTypeResource` instead of the
`MyEventTypeResource2`, which was obviously a joke.