aruha-1083 fix gzip flush
By default, Jetty output stream flushing behaves differently from Tomcat
when using compression. Calling flush would not cause the actual bytes
to be immediatelly flushed. This change in behaviour introduced a
problem, where subscription consumers would not get batches on time in
order to commit them. This way, their connections would be constantly
closed and they were not able to make progress.

With this change, all batches are flushed to the wire at the exact same
time as they used to be before switching to Jetty.