ARUHA-1079 Don't allow admins to override read and write permissions
Nakadi administrators should not be allowed to read from, or write to,
event types, if they are not in the list of readers or writers,
respectively.

It is still possible for admins to override the admin permissions
(this is necessary to help users who lock themselves out of their
event types), and add themselves to the list of readers and/or writers.