ARUHA-998 Validate admins before update
When updating the list of admins, check that all are valid.
It is possible that an existing admin isn't valid anymore (e.g., a
user existed when it was added to the list of admins, but does not
anymore when the list is updated). These will cause the update to fail:
it is good practice to remove deactivated admins, in case another
subject with the same name were created later.