aruha-1242 add event type and application to SLO log
By adding event type and the name of the "user" publishing, we are able
to easily identify top publishing applications using scalyr.

We already expose this data in our /metrics endpoint, but as we all know
very well, it has more than 2MB and zmon is just not able to handle it,
specially when running 400 instances for Black Friday.

On the other hand, scalyr, when properly used, is reliable and
performant.

I'm gonna use this new information to identify event types that would
require repartitioning for Black Friday.