Improve pull request template
This improve the PR template by moving the Zalando ticket ID out of
the PR title, yet still in a prominent position. It also makes it clear
that a ticket ID is not mandatory. With this template, we avoid giving
the false impression that only Zalando employees can contribute to
Nakadi.