ARUHA-1136 Warning header for event type creation and update
When creating or updating event types, it is possible to add a warning
header to the response. The purpose of this header is to warn users
that events in their event type may be read by archiving applications,
regardless of the authorization policy for the event type.

The warning header can be disabled by setting property
nakadi.authz.warnAllDataAccessMessage to an empty string.