Reduce logging by removing content duplication
On each requests to publish an event to Nakadi, 2 lines of logs were
generated: 1 for ACCESS_LOG and 1 for SLO log.

SLO log is used by scalyr to report SLO. It contains basically the same
information but with the addition of the request size in a more
structured way.

This change adds the body size to ACCESS_LOG and removes SLO for
publishing requests. This will result in an overall reduction of 30%,
which is the amount of logging we were requested to reduce in order to
avoid overspending on logs.