Remove partition key fields check
This feature was believed to be enabled for a long time now. It happens
that it was never enabled, since it breaks existen use cases. More
specifically event log, which has not predefined schema.

Since this feature was never there and now it was accidentally enabled,
this is a hotfix to completely remove.