Timeline with pre-existent topic
This change allows the creation of a timeline with a pre defined
topic. This is useful when the topic is created by some other tool, at
some other point in time.

One concrete example is KSQL, where the output of a query is written to
a topic, created by KSQL. Using this feature, the operator would be able
to expose this topic throw an event type, by creating a timeline and
pointing to this existent topic.