Add uniqueness constraint to original timeline table
This is to make sure this constraint is also present in the tests.

I made the acceptance test simpler by just using any string as a topic,
since we do not check if the topic actually exists. This is a "hack" for
administrators to use, so I considered that checking it is not strictly
required.