Cursor commit, patch and get protected by blacklisting CONSUMER_APP
In case a consumer application goes out of control and starts
hammering the cursors endpoints. It's important because these
endpoints are fairly expensive since they hit a non scalable
component: Zookeeper.

We consider that an app which is forbidden from consumption of events
is also forbidden from commits, which is part of the stream-commit
consumption process.