Detect changes to ArraySchema
This fixes a NPE changing an ArraySchema through generalization: an
array that used to define schemas for each item specifically now has a
single schema for all items.

We also added tests for the other way around: a schema that used to
have a single definition for all items now has a per item schema.