Introduce batch_timespan consumption parameter
In order to improve the SLO from Nakadi archiver, we are introducing a
new parameter that makes slicing events in 5 minutes window possible.

As it's explained in the description of the parameter, using
`batch_flush_timeout` alone is not enough and it lead to the
accumulation of events produced throughout a longer timespan than the
desired 5 minutes window.

To further illustrate when this happens consider the following
sequence of events:

1. events are produced one at a second starting at time 0
2. a stream to read these events is not openned until time 10. This
   stream has batch_flush_timeout of 5.
3. at time 15 a batch is flushed containing the events produced from
time 0 to time 15 (15 events).

This is not the desired behaviour for Nakadi archiver. Archiver would
like to get events from time 0 to time 5, another from time 5 to time
10 and finally a third from time 10 to time 15. Its goal is to slice
the topic by time.

This new parameter makes it possible.