Document how to choose number of partition of an event type
The section of the API responsible for defining the amount of partitions that an event type is given was elusive and unclear. It has caused pain for every use who ever tried to understand how many partitions they get when creating an event type. Even though Nakadi UI reduced this pain by providing a saner and more direct way to define the amount of partitions, the API was never reviewed and kept haunting new users with its elusive and unpractical definition.

This change in the documentation aims in making it more direct and plain honest.

We do plan on getting rid of this section of the API in the future, but while it doesn't happen, we hope this change will make the life of our users easier.