Allow for type narrowing under forward compatibility mode
Changing a schema from empty `{}` to something that is not empty
`{type: string}` should be seen as a type narrowing operation that
doesn't bread consumers, given the set of events accepted by the
narrowed definition is a subset of the original events. That means
that such a change should be allowed in this direction.

For fully compatible mode that is still not allowed as it means that
consumers developed based on the latest schema would not be able to
read events published under the oldest schema, contradicting the
definition of fully compatible.