[WIP] Upgrade to spring boot 2.3.1.RELEASE
This PR is not supposed to be merged and is here just to share the
current state that I find myself right now.

We did not upgrade spring boot since ever. This lead to accumulating a
lot of breaking changes.

My strategy was to get the project to build and boot by removing
dependencies that I couldn't reconciliate the version (due to
transitive incompatible dependencies) and lack of support for spring
boot 2 or spring framework 5+.

By no means I plan to remove all these features so please be
patient. Now that the project builds, I can start to bring them back
one by one. This way is better than try to upgrade everything at once,
which proved to be almost impossible task as the project didn't even
compile for days of work.