Fix failing test due to eventual consistency of default storage update
When updating the default storage one has to wait for the zk watcher
to be processed. This may take some time. In fact, we may have a bug
here when the watcher for some reason is missed (for example lost
connectivity to zk and recreation of watcher). But I'm not fixing
Nakadi, instead my goal is to first have more stable tests.

Also using teardown to avoid the failure of this particular test
cascade interfere with other tests which makes it harder to find the
"root cause" of the failures.