Upgrade Kafka Client to 2.5.0
Nakadi runs on ancient Kafka client 2.3.1 from year ago. Kafka is
preparing to replace Zookeeper by its own Raft based consensus
protocol and as part of this major step (expected to come with kafka 3
sometime in the future) they started to remove direct dependencies to
kafka, as for example topic creation, configuration and deletion.

The version of client that we used has an Admin that connected
directly to Kafka which is expected to stop to work sometime in the
future and worse than that exposed some internals from Kafka which
were never mean to be exposed.

The new admin client uses a new API added to kafka brokers and is
going to become the default one. It's also more powerful in the sense
that it allows to change broker configurations on the fly (even though
we don't need it for now).

Besides these Zookeeper related changes, the bump in version should
address several bug fixes, performance issues and unlock some more
features on the client.

On top of it all, by using spring to manage kafka client version, we
are more likely to upgrade it in the future, since spring manages its
version for us.

I'm not upgrading to 2.6.0 because this will happen when we upgrade
Spring boot (as soon as they close a new release). We desperately need
to bring some sanity back to our gradle files so I'm not overrideing
any versions now.

That all said, I also changed the kafka image used on tests by a new
kafka 2.6.0 which is expected to be released on production soon. It's
good to have tests dependencies running on the same version as
production software.